requirements.txt
blinker==1.9.0
cbor2==5.7.1
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
luma.core==2.5.3
luma.oled==3.14.0
MarkupSafe==3.0.3
pillow==12.0.0
psutil==7.1.3
RPi.GPIO==0.7.1
smbus2==0.5.0
Werkzeug==3.1.4
